requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.6
Jinja2==3.1.2
python-dateutil==2.8.2
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.4
matplotlib==3.7.2
seaborn==0.12.2
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pymysql==1.1.0 